Get all submissions from user
