+++ routes/assignments.service.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const dashboardService = require('../services/dashboard.service');
+const dashboardService = require('../services/assignments.service');
 
 const router = express.Router();
 
@@ -20,28 +20,28 @@ router.get('/segments/:id', async (req, res) => {
 
 // get user's 6 recent monitor log submissions
 // returns [ { drafts: [Submission, Submission, ...] } ]
-router.get('/recents/:id', async (req, res) => {
+router.get('/submissions/:id', async (req, res) => {
   const { id } = req.params;
   try {
-    const recentSubmissions = await dashboardService.getRecentSubmissions(id);
-    res.status(200).send(recentSubmissions);
+    const userSubmissions = await dashboardService.getUserSubmissions(id);
+    res.status(200).send(userSubmissions);
   } catch (err) {
     console.error(err);
     res.send(400).json({ message: err.message });
   }
 });
 
-// get user's unsubmitted log drafts
-// returns [ { submissions: [Submission, Submission, ...] } ]
-router.get('/unsubmitted/:id', async (req, res) => {
-  const { id } = req.params;
-  try {
-    const unsubmittedDrafts = await dashboardService.getUnsubmittedDrafts(id);
-    res.status(200).send(unsubmittedDrafts);
-  } catch (err) {
-    console.error(err);
-    res.send(400).json({ message: err.message });
-  }
-});
+// // get user's unsubmitted log drafts
+// // returns [ { submissions: [Submission, Submission, ...] } ]
+// router.get('/unsubmitted/:id', async (req, res) => {
+//   const { id } = req.params;
+//   try {
+//     const unsubmittedDrafts = await dashboardService.getUnsubmittedDrafts(id);
+//     res.status(200).send(unsubmittedDrafts);
+//   } catch (err) {
+//     console.error(err);
+//     res.send(400).json({ message: err.message });
+//   }
+// });
 
 module.exports = router;
